+++ .gitignore
@@ -7,7 +7,7 @@
 
 build/
 Pods/
-Podfile.lock
+*.lock
 *.pbxuser
 !default.pbxuser
 *.mode1v3
